requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
discord.py==2.4.0
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
aiosqlite==0.20.0
python-dotenv==1.0.1
requests==2.32.5
urllib3==2.6.2
Werkzeug==3.1.4
